--- apps/web/src/components/LiquidityChartRangeInput/Line.tsx
@@ -1,25 +1,0 @@
-import { ScaleLinear } from 'd3'
-import { useMemo } from 'react'
-import styled from 'styled-components'
-
-const StyledLine = styled.line`
-  opacity: 0.5;
-  stroke-width: 2;
-  stroke: ${({ theme }) => theme.colors.primary};
-  fill: none;
-  stroke-dasharray: 4;
-`
-
-export const Line = ({
-  value,
-  xScale,
-  innerHeight,
-}: {
-  value: number
-  xScale: ScaleLinear<number, number>
-  innerHeight: number
-}) =>
-  useMemo(
-    () => <StyledLine x1={xScale(value)} y1="0" x2={xScale(value)} y2={innerHeight} />,
-    [value, xScale, innerHeight],
-  )
