--- apps/web/src/components/LiquidityChartRangeInput/types.ts
@@ -1,90 +1,0 @@
-import { Bound } from 'config/constants/types'
-import { FeeAmount } from '@pancakeswap/v3-sdk'
-
-export interface ChartEntry {
-  activeLiquidity: number
-  price0: number
-}
-
-interface Dimensions {
-  width: number
-  height: number
-}
-
-interface Margins {
-  top: number
-  right: number
-  bottom: number
-  left: number
-}
-
-export interface ZoomLevels {
-  initialMin: number
-  initialMax: number
-  min: number
-  max: number
-}
-
-export interface LiquidityChartRangeInputProps {
-  // to distinguish between multiple charts in the DOM
-  id?: string
-
-  data: {
-    series: ChartEntry[]
-    current: number
-  }
-  ticksAtLimit: { [bound in Bound]?: boolean | undefined }
-
-  styles: {
-    area: {
-      // color of the ticks in range
-      selection: string
-    }
-
-    brush: {
-      handle: {
-        west: string
-        east: string
-      }
-    }
-  }
-
-  dimensions: Dimensions
-  margins: Margins
-
-  interactive?: boolean
-
-  brushLabels: (d: 'w' | 'e', x: number) => string
-  brushDomain: [number, number] | undefined
-  onBrushDomainChange: (domain: [number, number], mode: string | undefined) => void
-
-  zoomLevels: ZoomLevels
-  showZoomButtons?: boolean
-}
-
-export const ZOOM_LEVELS: Record<FeeAmount, ZoomLevels> = {
-  [FeeAmount.LOWEST]: {
-    initialMin: 0.999,
-    initialMax: 1.001,
-    min: 0.00001,
-    max: 1.5,
-  },
-  [FeeAmount.LOW]: {
-    initialMin: 0.999,
-    initialMax: 1.001,
-    min: 0.00001,
-    max: 1.5,
-  },
-  [FeeAmount.MEDIUM]: {
-    initialMin: 0.5,
-    initialMax: 2,
-    min: 0.00001,
-    max: 20,
-  },
-  [FeeAmount.HIGH]: {
-    initialMin: 0.5,
-    initialMax: 2,
-    min: 0.00001,
-    max: 20,
-  },
-}
