+++ apps/web/src/views/AddLiquidityV3/hooks/useDensityChartData.ts
@@ -1,10 +1,9 @@
 import { Currency } from '@pancakeswap/sdk'
 import { FeeAmount } from '@pancakeswap/v3-sdk'
+import { ChartEntry } from '@pancakeswap/uikit'
 import { TickProcessed } from 'hooks/v3/types'
 import { usePoolActiveLiquidity } from 'hooks/v3/usePoolTickData'
 import { useCallback, useMemo } from 'react'
-
-import { ChartEntry } from './types'
 
 export function useDensityChartData({
   currencyA,
