+++ apps/web/src/views/AddLiquidityV3/types.ts
@@ -1,5 +1,5 @@
 import { FeeAmount } from '@pancakeswap/v3-sdk'
-import { ZoomLevels, ZOOM_LEVELS } from 'components/LiquidityChartRangeInput/types'
+import { ZOOM_LEVELS, ZoomLevels } from '@pancakeswap/uikit'
 
 export enum SELECTOR_TYPE {
   V3,
