+++ packages/uikit/src/components/LiquidityChartRangeInput/types.ts
@@ -1,4 +1,5 @@
 import { BigintIsh } from "@pancakeswap/sdk";
+import { FeeAmount } from "@pancakeswap/v3-sdk";
 
 export interface ChartEntry {
   activeLiquidity: number;
@@ -79,3 +80,30 @@ export interface TickProcessed {
   liquidityNet: bigint;
   price0: string;
 }
+
+export const ZOOM_LEVELS: Record<FeeAmount, ZoomLevels> = {
+  [FeeAmount.LOWEST]: {
+    initialMin: 0.999,
+    initialMax: 1.001,
+    min: 0.00001,
+    max: 1.5,
+  },
+  [FeeAmount.LOW]: {
+    initialMin: 0.999,
+    initialMax: 1.001,
+    min: 0.00001,
+    max: 1.5,
+  },
+  [FeeAmount.MEDIUM]: {
+    initialMin: 0.5,
+    initialMax: 2,
+    min: 0.00001,
+    max: 20,
+  },
+  [FeeAmount.HIGH]: {
+    initialMin: 0.5,
+    initialMax: 2,
+    min: 0.00001,
+    max: 20,
+  },
+};
